changed ConnectionParams.Port to uint16 (match limit 65,535)

+++ pkg/core/driver.go
@@ -3,7 +3,7 @@ package core
 type ConnectionParams struct {
 	Host     string
 	Instance string
-	Port     uint
+	Port     uint16
 	User     string
 	Password string
 	Database string
